requirements.txt
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
bmi-calculator==0.0.2
boto==2.49.0
boto3==1.16.24
botocore==1.19.24
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==2.1.1
django-allauth==0.43.0
django-crispy-forms==1.10.0
django-storages==1.9.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
